Flatlist removed from upcoming leaves
