+++ src/app/screens/HomeScreen/teamMembersLeaves/index.tsx
@@ -1,39 +1,10 @@
-import {
-  ActivityIndicator,
-  FlatList,
-  ListRenderItemInfo,
-  StyleSheet,
-  View,
-} from 'react-native';
-import React from 'react';
+import {ActivityIndicator, StyleSheet, View} from 'react-native';
+import React, {Fragment} from 'react';
 
-import linear from '../../../components/seperator/linear';
+import Linear from '../../../components/seperator/linear';
+import LeaveCard from './leaveCard';
 import Typography from '../../../components/typography';
 import {useTeamMembersLeaves} from '../dashboard.hooks';
-
-import {dateFormate} from '../../../utils/date';
-
-type ItemProps = {
-  name: string;
-  from: string;
-  to: string;
-  days: number;
-};
-
-const renderItem = ({
-  item: {name, from, to, days},
-}: ListRenderItemInfo<ItemProps>) => {
-  return (
-    <View style={styles.item}>
-      <Typography type="header">{name}</Typography>
-      <Typography type="label">
-        {days === 1
-          ? `Leave on ${dateFormate(from)}`
-          : `Leave From: ${dateFormate(from)} To: ${dateFormate(to)}`}
-      </Typography>
-    </View>
-  );
-};
 
 const TeamMembersLeaves = () => {
   const {data, isLoading} = useTeamMembersLeaves();
@@ -50,11 +21,12 @@ const TeamMembersLeaves = () => {
       {isLoading ? (
         <ActivityIndicator size="large" />
       ) : (
-        <FlatList
-          data={data}
-          renderItem={renderItem}
-          ItemSeparatorComponent={linear}
-        />
+        data.map((item, index) => (
+          <Fragment key={index}>
+            {Boolean(index) && <Linear />}
+            <LeaveCard {...item} />
+          </Fragment>
+        ))
       )}
     </View>
   );
@@ -63,15 +35,11 @@ const TeamMembersLeaves = () => {
 const styles = StyleSheet.create({
   container: {
     padding: 16,
+    marginBottom: 72,
   },
   title: {
     fontSize: 16,
     fontWeight: '700',
-    paddingVertical: 12,
-  },
-  item: {
-    flexDirection: 'column',
-    gap: 6,
     paddingVertical: 12,
   },
 });
